sharepoint-ai-assistant/requirements.txt
streamlit
langchain>=0.1.0
office365-rest-python-client
pandas
ollama
python-docx
pdfplumber
openpyxl
requests
